RFC-4465, Implement step 2, moving functions to different files to mirror cuda path
