+++ tests/v1/test_torch_ops.py
@@ -3,6 +3,7 @@
 from typing import Any, Union
 import ctypes
 import os
+import sys
 import time
 import unittest.mock
 
@@ -18,6 +19,8 @@ from lmcache.v1.multiprocess.native_completion import (
 from lmcache.v1.platform import resolve_device_ops
 from lmcache.v1.platform import torch_ops as _py_ops
 import lmcache.lmcache_native as lmcache_native
+
+tensor_from_ptr = sys.modules["lmcache.v1.platform.torch_ops._tensor_from_ptr"]
 
 # ==========================================
 # 0. utils functions.
@@ -3081,8 +3084,8 @@ def test_tensor_from_ptr_routes_musa_pointer(
         )
         return torch.empty(shape, dtype=dtype)
 
-    monkeypatch.setattr(_py_ops.torch, "device", FakeDevice)
-    monkeypatch.setattr(_py_ops, "_tensor_from_musa_ptr", fake_musa_ptr)
+    monkeypatch.setattr(tensor_from_ptr.torch, "device", FakeDevice)
+    monkeypatch.setattr(tensor_from_ptr, "_tensor_from_musa_ptr", fake_musa_ptr)
 
     tensor = _py_ops._tensor_from_ptr(0x1000, (2, 3), torch.float16, "musa:0")
 
@@ -3142,12 +3145,12 @@ def test_tensor_from_musa_ptr_uses_external_storage(
         return fake_tensor
 
     monkeypatch.setattr(
-        _py_ops.torch._C,
+        tensor_from_ptr.torch._C,
         "_construct_storage_from_data_pointer",
         fake_construct_storage,
         raising=False,
     )
-    monkeypatch.setattr(_py_ops.torch, "empty", fake_empty)
+    monkeypatch.setattr(tensor_from_ptr.torch, "empty", fake_empty)
 
     result = _py_ops._tensor_from_musa_ptr(
         0x1000, (2, 3), torch.float16, fake_device, 12
@@ -3179,7 +3182,7 @@ def test_tensor_from_musa_ptr_fails_without_external_storage(
         raise RuntimeError("storage construction unavailable")
 
     monkeypatch.setattr(
-        _py_ops.torch._C,
+        tensor_from_ptr.torch._C,
         "_construct_storage_from_data_pointer",
         fake_construct_storage,
         raising=False,
